BINIUS-556: Split the hash crate along the prover/verifier line
The SIMD kernels and the threaded Merkle tree construction sat in the same
crate a verifier links against, so reading the verifier's trust base meant
reading past code it never calls.

binius-hash now holds only the sequential side: a leaf hash, a two-to-one
compression, serialization, and the suite trait that pairs them. It contains
no vector kernel, no thread pool, and denies unsafe code outright, with the
one required exception marked where it stands.

binius-hash-prover holds the batched side: the multi-lane SHA-256 and Blake3
kernels, the parallel digest and compression traits, and the Merkle tree
builder.

The suite trait splits with the crates. The verifier's names the sequential
leaf hash and compression. The prover's extends it with their batched
counterparts, so prover-side generics keep reading H::LeafHash unchanged and
only the bound and the import move.

Every batched path stays pinned byte-for-byte to the sequential one it stands
in for, which is what lets both sides commit to the same tree.
