+++ crates/hash-prover/Cargo.toml
@@ -1,5 +1,7 @@
+# Copyright 2026 The Binius Developers
+
 [package]
-name = "binius-iop-prover"
+name = "binius-hash-prover"
 version.workspace = true
 edition.workspace = true
 authors.workspace = true
@@ -9,33 +11,31 @@ workspace = true
 
 [dependencies]
 binius-compute = { path = "../compute" }
-binius-core = { path = "../core" }
 binius-field = { path = "../field" }
 binius-hash = { path = "../hash" }
-binius-hash-prover = { path = "../hash-prover" }
-binius-iop = { path = "../iop" }
-binius-ip = { path = "../ip" }
-binius-ip-prover = { path = "../ip-prover" }
-binius-math = { path = "../math" }
-binius-transcript = { path = "../transcript" }
 binius-utils = { path = "../utils" }
+blake3 = { workspace = true, features = ["traits-preview"] }
+bytemuck.workspace = true
+bytes.workspace = true
 digest.workspace = true
-getset.workspace = true
 itertools.workspace = true
 rand.workspace = true
+sha2.workspace = true
+thiserror.workspace = true
 tracing.workspace = true
 
 [dev-dependencies]
-anyhow.workspace = true
-binius-math = { path = "../math", features = ["test-utils"] }
 criterion.workspace = true
 proptest.workspace = true
 rand = { workspace = true, features = ["thread_rng"] }
-sha2.workspace = true
 
 [features]
 default = []
 
 [[bench]]
-name = "binary_merkle_tree"
+name = "sha256"
+harness = false
+
+[[bench]]
+name = "blake3_portable"
 harness = false
