+++ crates/recursion/Cargo.toml
@@ -13,6 +13,7 @@ binius-core = { path = "../core" }
 binius-field = { path = "../field" }
 binius-frontend = { path = "../frontend" }
 binius-hash = { path = "../hash" }
+binius-hash-prover = { path = "../hash-prover" }
 binius-iop = { path = "../iop" }
 binius-ip = { path = "../ip" }
 binius-transcript = { path = "../transcript" }
